% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_report_22.R
\name{aggregate_countries}
\alias{aggregate_countries}
\title{Produce aggregate and long form, ready for plot}
\usage{
aggregate_countries(dt, aggregation_level)
}
\arguments{
\item{dt}{data.table, all countries, wide form}

\item{aggregation_level}{from user input}
}
\description{
Function takes wide form, with all the countries, then
- aggregates countries
- produces long form
}
