% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aws.R
\name{get_aws_connection}
\alias{get_aws_connection}
\title{Basic function to establish connection with aws, reading parameters from the config.yml}
\usage{
get_aws_connection()
}
\description{
Basic function to establish connection with aws, reading parameters from the config.yml
}
