% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aws.R
\name{get_clean_leads_profiles}
\alias{get_clean_leads_profiles}
\title{Return leads+profiles for dashboard charting}
\usage{
get_clean_leads_profiles()
}
\description{
Reading clean profiles using [read_leads_profiles_clean()] function, and then adjusting column types
}
