% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aws.R
\name{get_clean_profiles}
\alias{get_clean_profiles}
\title{Return profiles for dashboard charting}
\usage{
get_clean_profiles()
}
\description{
Reading clean profiles using [read_clean_profiles()] function, and then adjusting column types
}
