% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aws.R
\name{get_credited_leads}
\alias{get_credited_leads}
\title{Return only credited leads}
\usage{
get_credited_leads(dt)
}
\arguments{
\item{dt}{input data.table, after [mark_credited_leads()]}
}
\description{
return only credited leads, the function is a next step after [mark_credited_leads()]
}
