% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aws.R
\name{get_default_analysis_period}
\alias{get_default_analysis_period}
\title{Generate default analysis period}
\usage{
get_default_analysis_period()
}
\description{
Generate default period, which is full last month, and current month including today()
}
