% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aws.R
\name{get_lookup_table}
\alias{get_lookup_table}
\title{Get lookup_dt table}
\usage{
get_lookup_table()
}
\description{
Function to
- read lookup_dt table from aws
- adjust Encoding to latin1
- change column names
}
