% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aws.R
\name{get_primary_industry_table}
\alias{get_primary_industry_table}
\title{Get Primary Industry table}
\usage{
get_primary_industry_table()
}
\description{
Function performs
- reads table from aws, using [read_primary_industry_table()] function
- adjusts Encoding to latin1
- formats column types
}
