% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aws.R
\name{get_profiles_investment_daily}
\alias{get_profiles_investment_daily}
\title{Get active profiles, per each day, with a minimum_investment per each day}
\usage{
get_profiles_investment_daily(dt, analysis_period)
}
\description{
Function is very similar to [get_active_profiles_daily()], but it also takes into account changes of minimum investment levels
}
