% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aws.R
\name{mark_credited_leads}
\alias{mark_credited_leads}
\title{Identify credited leads}
\usage{
mark_credited_leads(dt)
}
\arguments{
\item{dt}{input data.table}
}
\description{
Function checks leads statuses, picking Sent and Credited
}
\details{
Function logic has following steps
- reduce changes to only these, having changed status, "flatten" all other changes, when status is not changed
- get lead_id, which were credited, sometimes a couple of days later, after initial send
- subset lead_id to these, which have been sent
- perform two joins of these datasets - leads_sent and credited
- as a result, we have a dataset with the earliest "sent" status, and also a flag credited = y/n (crediting occurs a few days after lead is sent)
}
