% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_report_22.R
\name{produce_daily_profile_plot}
\alias{produce_daily_profile_plot}
\title{Produce plot with sessions or transactions}
\usage{
produce_daily_profile_plot(dt, metric_selection, time_selection)
}
\arguments{
\item{dt}{data.table with input value, long form}

\item{metric}{whether we should plot sessions or transactions}
}
\description{
Produce plot
}
