% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_report_21.R
\name{produce_join_table_for_industries}
\alias{produce_join_table_for_industries}
\title{Join tables for Industries report}
\usage{
produce_join_table_for_industries()
}
\description{
Function which wraps components and produces table for the Industries report
}
