% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aws.R
\name{read_clean_leads}
\alias{read_clean_leads}
\title{Reading subset of columns from a leads database in aws}
\usage{
read_clean_leads()
}
\description{
Reading subset of columns from a leads database in aws
}
