% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aws.R
\name{read_clean_profiles}
\alias{read_clean_profiles}
\title{Reading subset of columns from a profiles table in aws}
\usage{
read_clean_profiles()
}
\description{
Reading subset of columns from a profiles table in aws
}
