% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aws.R
\name{read_table_from_aws}
\alias{read_table_from_aws}
\title{Reading whole table from aws}
\usage{
read_table_from_aws(table_name)
}
\arguments{
\item{table_name}{name of the table to read}
}
\description{
Reading whole table from aws
}
