% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aws.R
\name{write_table_to_aws}
\alias{write_table_to_aws}
\title{Writing to a table on aws}
\usage{
write_table_to_aws(dt, table_name, overwrite = TRUE, append = FALSE)
}
\arguments{
\item{dt}{data.table, to be stored in aws database}

\item{table_name}{name of the table, where we will store data}

\item{overwrite}{passing a parameter for RMySQL::dbWriteTable; it is set TRUE as default}

\item{append}{passing a parameter for RMySQL::dbWriteTable; it is set FALSE as default}
}
\description{
Writing to a table on aws
}
